Move ladder deletion from ladder show page to ladders management page.
