+++ app/routes/games/ladders-manage.js
@@ -1,6 +1,8 @@
+import { action } from '@ember/object';
 import Route from '@ember/routing/route';
 import { inject as service } from '@ember/service';
 import RSVP from 'rsvp';
+
 
 export default class GamesLaddersRoute extends Route {
   @service store;
@@ -13,5 +15,10 @@ export default class GamesLaddersRoute extends Route {
       sideLadders: this.store.query(
         'ladder', {game_id: params.game_id, kind: 'side'})
     });
+  }
+
+  @action
+  refreshModel() {
+    this.refresh();
   }
 }
