+++ app/templates/games/ladder-new.hbs
@@ -1,6 +1,6 @@
 <h1>
   Create ladder for:
-  <LinkTo @route="games.ladders" @model={{@model.game.id}}>
+  <LinkTo @route="games.ladders-manage" @model={{@model.game.id}}>
     {{@model.game.name}}
   </LinkTo>
 </h1>
