+++ app/templates/games/ladders-manage.hbs
@@ -1,0 +1,36 @@
+<h1>
+  <LinkTo @route="games.show" @model={{@model.game.id}}>
+    {{@model.game.name}}
+  </LinkTo>
+  - Manage Ladders
+</h1>
+
+<h2>Main ladders</h2>
+
+<div class="main-ladders">
+  {{#if @model.mainLadders}}
+    <LaddersManageCategory
+     @ladders={{@model.mainLadders}}
+     @deleteLadder={{this.deleteLadder}} />
+  {{else}}
+    (No main ladders)
+  {{/if}}
+</div>
+
+<h2>Side ladders</h2>
+
+<div class="side-ladders">
+  {{#if @model.sideLadders}}
+    <LaddersManageCategory
+     @ladders={{@model.sideLadders}}
+     @deleteLadder={{this.deleteLadder}} />
+  {{else}}
+    (No side ladders)
+  {{/if}}
+</div>
+
+<div class="error-message" id="ladder-delete-error"></div>
+
+<LinkTo @route="games.ladder-new" @model={{@model.game.id}}>
+  Create ladder
+</LinkTo>
