+++ app/templates/games/show.hbs
@@ -32,7 +32,7 @@
     <LinkTo @route="games.filter-groups" @model={{@model.game.id}}> Filter groups admin </LinkTo>
   </li>
   <li>
-    <LinkTo @route="games.ladders" @model={{@model.game.id}}> Ladders admin </LinkTo>
+    <LinkTo @route="games.ladders-manage" @model={{@model.game.id}}> Ladders admin </LinkTo>
   </li>
 </ul>
